+++ src/resources/users/user.router.js
@@ -3,7 +3,7 @@
  */
 
 const router = require('express').Router();
-const User = require('./user.model');
+const User = require('src/resources/users/user.model');
 const usersService = require('./user.service');
 
 /**
